# docs/openapi.yaml
openapi: 3.1.0
info:
  title: GUM Internships API
  version: 1.0.0
  description: >
    MVP API contract. Every response uses the envelope {success, data, error, meta}.
    DB is snake_case; this API is camelCase. Auth = Bearer JWT (access 15m, refresh 30d).
    Pagination on list endpoints: ?page=&limit= (default 20, max 100).
    Endpoints marked "Public" need no token. x-required-roles lists allowed roles.
servers:
  - url: https://api.gum-internships.example.com/v1
    description: Production
  - url: https://staging-api.gum-internships.example.com/v1
    description: Staging
security:
  - bearerAuth: []
...
paths:
...
  /lessons/{lessonId}/video-upload:
    post:
      tags: [Internships]
      summary: Create Bunny Stream video + return TUS upload credentials
      x-required-roles: [instructor, moderator, super_admin]
      parameters:
        - { name: lessonId, in: path, required: true, schema: { type: integer, format: int64 } }
      responses:
        '200':
          description: Upload target
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - properties:
                      data:
                        type: object
                        properties:
                          bunnyVideoId: { type: string }
                          tusEndpoint: { type: string }
                          authorizationSignature: { type: string }
                          authorizationExpire: { type: integer }
                          libraryId: { type: string }

  # ============================================================ BATCHES
  /internships/{internshipId}/batches:
    parameters:
      - { name: internshipId, in: path, required: true, schema: { type: integer, format: int64 } }
    get:
      tags: [Batches]
      summary: List batches of an internship
      x-required-roles: [instructor, moderator, super_admin]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - properties:
                      data: { type: array, items: { $ref: '#/components/schemas/Batch' } }
    post:
      tags: [Batches]
      summary: Create batch
      x-required-roles: [instructor, moderator, super_admin]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/BatchCreate' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - properties: { data: { $ref: '#/components/schemas/Batch' } }
  /batches/{batchId}:
    patch:
      tags: [Batches]
      summary: Update batch (seats, dates, status)
      x-required-roles: [instructor, moderator, super_admin]
      parameters:
        - { name: batchId, in: path, required: true, schema: { type: integer, format: int64 } }
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/BatchCreate' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - properties: { data: { $ref: '#/components/schemas/Batch' } }
  /batches/{batchId}/students:
    get:
      tags: [Batches]
      summary: Enrolled students with progress & attendance % (own batches)
      x-required-roles: [instructor, moderator, super_admin]
      parameters:
        - { name: batchId, in: path, required: true, schema: { type: integer, format: int64 } }
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - properties:
                      data:
                        type: array
                        items:
                          type: object
                          properties:
                            enrollmentId: { type: integer, format: int64 }
                            studentName: { type: string }
                            progressPercent: { type: number }
                            attendancePercent: { type: number }
                            status: { type: string }
                      meta: { $ref: '#/components/schemas/PaginationMeta' }

  # ============================================================ ENROLLMENTS
...
  /lessons/{lessonId}/play:
    get:
      tags: [Learning]
      summary: Signed Bunny Stream playback URL (active enrollment or preview lesson)
      x-required-roles: [student]
      parameters:
        - { name: lessonId, in: path, required: true, schema: { type: integer, format: int64 } }
        - { name: enrollmentId, in: query, required: true, schema: { type: integer, format: int64 } }
      responses:
        '200':
          description: Signed playback
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - properties:
                      data:
                        type: object
                        properties:
                          playbackUrl: { type: string }
                          expiresAt: { type: string, format: date-time }
        '403': { $ref: '#/components/responses/Forbidden' }
  /lessons/{lessonId}/progress:
    post:
      tags: [Learning]
      summary: Report watch progress / completion (recomputes enrollment %)
      x-required-roles: [student]
      parameters:
        - { name: lessonId, in: path, required: true, schema: { type: integer, format: int64 } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [enrollmentId]
              properties:
                enrollmentId: { type: integer, format: int64 }
                watchedSeconds: { type: integer, minimum: 0 }
                completed: { type: boolean }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - properties:
                      data:
                        type: object
                        properties:
                          lessonProgress: { $ref: '#/components/schemas/LessonProgress' }
                          progressPercent: { type: number }

  # ============================================================ LIVE SESSIONS
...
  /orders:
    post:
      tags: [Commerce]
      summary: Create order for a PAID internship (server computes GST; returns Razorpay checkout params) — WEB ONLY
      x-required-roles: [student]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/OrderCreate' }
      responses:
        '201':
          description: Order + checkout params
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - properties: { data: { $ref: '#/components/schemas/CheckoutParams' } }
        '409': { $ref: '#/components/responses/Conflict' }
  /orders/me:
    get:
      tags: [Commerce]
      summary: My orders
      x-required-roles: [student]
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - properties:
                      data: { type: array, items: { $ref: '#/components/schemas/Order' } }
                      meta: { $ref: '#/components/schemas/PaginationMeta' }
  /orders/{orderId}:
    get:
      tags: [Commerce]
      summary: Order detail (owner only)
      x-required-roles: [student]
      parameters:
        - { name: orderId, in: path, required: true, schema: { type: integer, format: int64 } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - properties: { data: { $ref: '#/components/schemas/Order' } }
  /orders/{orderId}/invoice:
    get:
      tags: [Commerce]
      summary: Signed URL for GST invoice PDF
      x-required-roles: [student]
      parameters:
        - { name: orderId, in: path, required: true, schema: { type: integer, format: int64 } }
      responses:
        '200': { $ref: '#/components/responses/UploadUrlResponse' }
  /orders/{orderId}/refund-request:
    post:
      tags: [Commerce]
      summary: Request refund (goes to finance queue)
      x-required-roles: [student]
      parameters:
        - { name: orderId, in: path, required: true, schema: { type: integer, format: int64 } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason: { type: string, minLength: 10 }
      responses:
        '201':
          description: Requested
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - properties: { data: { $ref: '#/components/schemas/Refund' } }
  /coupons/validate:
    post:
      tags: [Commerce]
      summary: Preview coupon discount for an internship
      x-required-roles: [student]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code, internshipId]
              properties:
                code: { type: string }
                internshipId: { type: integer, format: int64 }
      responses:
        '200':
          description: Validation result (valid=false carries reason, still 200)
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - properties: { data: { $ref: '#/components/schemas/CouponValidation' } }
  /payments/razorpay/webhook:
    post:
      tags: [Commerce]
      summary: 'Public: Razorpay webhook (HMAC-verified, idempotent). Events: payment.captured, payment.failed, refund.processed'
      security: []
      parameters:
        - { name: X-Razorpay-Signature, in: header, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, description: Raw Razorpay event payload }
      responses:
        '200': { description: 'Acknowledged (always 200 for valid signatures, even on no-op replays)' }
        '401': { description: Bad signature }

  # ============================================================ PROJECTS & SUBMISSIONS
...
  /admin/refunds:
    get:
      tags: [Admin]
      summary: Refund queue (default status=requested)
      x-required-roles: [finance_admin, super_admin]
      parameters:
        - { name: status, in: query, schema: { type: string, enum: [requested, approved, rejected, processed] } }
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - properties:
                      data: { type: array, items: { $ref: '#/components/schemas/Refund' } }
                      meta: { $ref: '#/components/schemas/PaginationMeta' }
  /admin/refunds/{refundId}/decision:
    post:
      tags: [Admin]
      summary: Approve (triggers Razorpay refund) or reject
      x-required-roles: [finance_admin, super_admin]
      parameters:
        - { name: refundId, in: path, required: true, schema: { type: integer, format: int64 } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [decision]
              properties:
                decision: { type: string, enum: [approved, rejected] }
                reason: { type: string, description: Required when rejected }
      responses:
        '200': { $ref: '#/components/responses/OkMessage' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
  responses:
    OkMessage:
      description: OK
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/Envelope'
              - properties:
                  data:
                    type: object
                    properties: { message: { type: string } }
...
    UploadUrlResponse:
      description: Presigned upload target
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/Envelope'
              - properties:
                  data:
                    type: object
                    properties:
                      uploadUrl: { type: string }
                      publicUrl: { type: string }
                      headers: { type: object, additionalProperties: { type: string } }
                      expiresAt: { type: string, format: date-time }
...
    Forbidden:
      description: 'Role/ownership check failed (error.code=FORBIDDEN)'
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
    Conflict:
      description: 'State conflict, e.g. ALREADY_ENROLLED, BATCH_FULL, SUBMISSION_PENDING_REVIEW'
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
  schemas:
    Envelope:
      type: object
      required: [success, data, error]
      properties:
        success: { type: boolean, const: true }
        data: {}
        error: { type: 'null' }
        meta: { type: object }
    ErrorEnvelope:
      type: object
      required: [success, data, error]
      properties:
        success: { type: boolean, const: false }
        data: { type: 'null' }
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string, description: 'Stable machine code, e.g. VALIDATION_ERROR, FORBIDDEN, BATCH_FULL' }
            message: { type: string }
            details: { description: Optional zod issue list }
    PaginationMeta:
      type: object
      properties:
        pagination:
          type: object
          properties:
            page: { type: integer }
            limit: { type: integer }
            total: { type: integer }
            totalPages: { type: integer }
...
    Batch:
      type: object
      properties:
        id: { type: integer, format: int64 }
        internshipId: { type: integer, format: int64 }
        name: { type: string }
        startDate: { type: string, format: date }
        endDate: { type: string, format: date }
        enrollmentDeadline: { type: [string, 'null'], format: date-time }
        seatsTotal: { type: integer }
        seatsFilled: { type: integer }
        seatsLeft: { type: integer }
        waitlistEnabled: { type: boolean }
        status: { type: string, enum: [scheduled, enrolling, ongoing, completed, cancelled] }
    BatchCreate:
      type: object
      required: [name, startDate, endDate, seatsTotal]
      properties:
        name: { type: string }
        startDate: { type: string, format: date }
        endDate: { type: string, format: date }
        enrollmentDeadline: { type: string, format: date-time }
        seatsTotal: { type: integer, minimum: 1 }
        waitlistEnabled: { type: boolean, default: false }
        waitlistLimit: { type: integer, minimum: 1 }
...
    LessonProgress:
      type: object
      properties:
        lessonId: { type: integer, format: int64 }
        status: { type: string, enum: [not_started, in_progress, completed] }
        watchedSeconds: { type: integer }
        completedAt: { type: [string, 'null'], format: date-time }
    Order:
      type: object
      properties:
        id: { type: integer, format: int64 }
        orderNo: { type: string }
        internshipId: { type: integer, format: int64 }
        batchId: { type: [integer, 'null'], format: int64 }
        subtotal: { type: number }
        discountAmount: { type: number }
        taxableAmount: { type: number }
        gstRate: { type: number }
        gstAmount: { type: number }
        totalAmount: { type: number }
        currency: { type: string }
        status: { type: string, enum: [created, pending, paid, failed, refunded, cancelled] }
        razorpayOrderId: { type: [string, 'null'] }
        invoiceNo: { type: [string, 'null'] }
        createdAt: { type: string, format: date-time }
    OrderCreate:
      type: object
      required: [internshipId, billingName, billingEmail, billingPhone, billingState]
      properties:
        internshipId: { type: integer, format: int64 }
        batchId: { type: integer, format: int64 }
        couponCode: { type: string }
        billingName: { type: string }
        billingEmail: { type: string, format: email }
        billingPhone: { type: string }
        billingState: { type: string, description: GST place of supply }
        billingGstin: { type: string }
    CheckoutParams:
      type: object
      description: Params for Razorpay Checkout.js on web
      properties:
        order: { $ref: '#/components/schemas/Order' }
        razorpayKeyId: { type: string }
        razorpayOrderId: { type: string }
        amountPaise: { type: integer }
        currency: { type: string }
        prefill: { type: object, properties: { name: { type: string }, email: { type: string }, contact: { type: string } } }
    CouponValidation:
      type: object
      properties:
        valid: { type: boolean }
        code: { type: string }
        discountAmount: { type: number }
        finalTotal: { type: number }
        reason: { type: [string, 'null'], description: Why invalid, when valid=false }
    Refund:
      type: object
      properties:
        id: { type: integer, format: int64 }
        orderId: { type: integer, format: int64 }
        amount: { type: number }
        reason: { type: [string, 'null'] }
        status: { type: string, enum: [requested, approved, rejected, processed] }
        createdAt: { type: string, format: date-time }
...